Sketch out reconciliation logic for Role and RoleBinding
The main shortcuts taken are around efficiency and consistency.

This currently relies on reseting the entire state of the graph for EVERY update, deleting everything and adding. This will result in lots of extra snapshots in the database, and therefore excessive index sizes and maintenance.

In reality, we can very likely do better by:

1. Retrieving current state from inventory (with a lock ID)
2. Diffing against the latest state in Kubernetes
3. Making the necessary updates directly in Inventory (with a lock ID)

If the lock ID no longer matches, we abort (it means we have lost our leader lease).

If we streamed individual updates from the cluster, we would be able to do improve step 2 greatly.